Move redux actions to dispatchers folder
